mini-siem-agent/requirements_win_agent.txt
requests
PyYAML
pywin32